requirements.txt
numpy==1.21.5
Pillow==9.0.0
torch==1.10.1+cu102
torchaudio==0.10.1+cu102
torchvision==0.11.2+cu102
typing-extensions==4.0.1
